base: Actually fix the padding on StatusBar
Android P gets a new QS layout that looks good
but it has some issues on smaller screens devices,
like the rounded buttons clipping with the borders
and the date not looking really good.

LineageOS suggested a fix (https://review.lineageos.org/224021)
but it's circumventing it because it adds padding and
some people doesn't like the padding, so, instead
let's apply the 8dp spacing over the affected headers
and the small qs tile.

Signed-off-by: Rohit Vashisht <rohitv88cf@gmail.com>